Refactor details.py by splitting into details and comment file, moving helper methods, renaming maps
